requirements.txt
# NukkadMart Backend Dependencies

# FastAPI Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Pydantic for data validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database Drivers
motor==3.3.2           # Async MongoDB driver
pymongo==4.6.1         # MongoDB driver
# Using Upstash Redis REST API via httpx (no redis package needed)

# AWS SDK for Bedrock/Nova
boto3>=1.34.34
botocore>=1.34.34

# AWS Nova Sonic Bidirectional Streaming SDK (requires Python 3.12+)
aws-sdk-bedrock-runtime>=0.4.0

# AI Services
groq==0.4.2            # Groq API client

# Image Processing
Pillow==10.2.0


# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# WebSocket Support
websockets==12.0

# Utilities
python-dotenv==1.0.1
orjson==3.9.13         # Fast JSON serialization

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
